Detach Diff from RxDataSources, and cleanup
:zap: WIP

Remove binaries
