+++ Sources/Differentiator/SectionModelType.swift
