+++ Sources/Differentiator/String+IdentifiableType.swift
