+++ Sources/RxDataSources/DataSources+Rx/RxTableViewSectionedAnimatedDataSource.swift
@@ -12,6 +12,7 @@ import UIKit
 import RxSwift
 import RxCocoa
 #endif
+import Differentiator
 
 open class RxTableViewSectionedAnimatedDataSource<S: AnimatableSectionModelType>
     : TableViewSectionedDataSource<S>
@@ -46,7 +47,7 @@ open class RxTableViewSectionedAnimatedDataSource<S: AnimatableSectionModelType>
                     }
                     let oldSections = dataSource.sectionModels
                     do {
-                        let differences = try differencesForSectionedView(initialSections: oldSections, finalSections: newSections)
+                        let differences = try Diff.differencesForSectionedView(initialSections: oldSections, finalSections: newSections)
 
                         for difference in differences {
                             dataSource.setSections(difference.finalSections)
