+++ Sources/RxDataSources/DataSources/CollectionViewSectionedDataSource.swift
@@ -11,6 +11,7 @@ import UIKit
 #if !RX_NO_MODULE
 import RxCocoa
 #endif
+import Differentiator
     
 open class _CollectionViewSectionedDataSource
     : NSObject
