+++ Sources/RxDataSources/DataSources/DataSources.swift
@@ -8,28 +8,29 @@
 
 import Foundation
 
+@_exported import Differentiator
+
 enum RxDataSourceError : Error {
-    case unwrappingOptional
-    case preconditionFailed(message: String)
+  case preconditionFailed(message: String)
 }
 
 func rxPrecondition(_ condition: Bool, _ message: @autoclosure() -> String) throws -> () {
-    if condition {
-        return
-    }
-    rxDebugFatalError("Precondition failed")
+  if condition {
+    return
+  }
+  rxDebugFatalError("Precondition failed")
 
-    throw RxDataSourceError.preconditionFailed(message: message())
+  throw RxDataSourceError.preconditionFailed(message: message())
 }
 
 func rxDebugFatalError(_ error: Error) {
-    rxDebugFatalError("\(error)")
+  rxDebugFatalError("\(error)")
 }
 
 func rxDebugFatalError(_ message: String) {
-    #if DEBUG
-        fatalError(message)
-    #else
-        print(message)
-    #endif
+  #if DEBUG
+    fatalError(message)
+  #else
+    print(message)
+  #endif
 }
