+++ Sources/RxDataSources/DataSources/UI+SectionedViewType.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 import UIKit
+import Differentiator
 
 func indexSet(_ values: [Int]) -> IndexSet {
     let indexSet = NSMutableIndexSet()
