+++ Tests/ChangeSet+TestExtensions.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import Differentiator
 import RxDataSources
 
 fileprivate class ItemModelTypeWrapper<I> {
