+++ Tests/Randomizer.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import Differentiator
 import RxDataSources
 
 // https://en.wikipedia.org/wiki/Random_number_generation
