+++ Tests/XCTest+Extensions.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 import XCTest
+import Differentiator
 import RxDataSources
 
 func XCAssertEqual<S: AnimatableSectionModelType>(_ lhs: [S], _ rhs: [S], file: StaticString = #file, line: UInt = #line)
