+++ Tests/s.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import Differentiator
 import RxDataSources
 
 /**
